Revert "refactor(gemini): centralize native error handling"
This reverts commit a9f0b30436f5bd3b1af00688df671adc0f606d7d.
